update to use typescript and npm
