+++ .gitignore
@@ -164,14 +164,7 @@ docs/build
 
 # Log files
 npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
+npm-error.log*
 
-# Yarn leftovers
-.yarn/*
-!.yarn/patches
-!.yarn/releases
-!.yarn/plugins
-!.yarn/sdks
-!.yarn/versions
-.pnp.*
+# npm leftovers
+.npm/
