+++ .gitpod.yml
@@ -1,6 +1,6 @@
 tasks:
-  - init: cd docs && yarn install
-    command: yarn start
+  - init: cd docs && npm install
+    command: npm start
 ports:
   - port: 3000
     onOpen: open-preview
