+++ docs/.husky/pre-commit
@@ -2,4 +2,4 @@
 . "$(dirname -- "$0")/_/husky.sh"
 
 cd docs
-yarn run lint-staged
+npm run lint-staged
