+++ docs/src/theme/NotFound/Content/index.tsx
@@ -2,7 +2,11 @@ import React from 'react';
 import clsx from 'clsx';
 import Heading from '@theme/Heading';
 
-export default function NotFoundContent({className}) {
+interface Props {
+  className?: string;
+}
+
+export default function NotFoundContent({ className }: Props) {
   return (
     <main className={clsx('container margin-vert--xl', className)}>
       <div className="row">
@@ -10,7 +14,7 @@ export default function NotFoundContent({className}) {
           <Heading as="h1" className="hero__title">
             Page not found
           </Heading>
-          <p>That URL doesn't exist on this site. The docs were recently reorganized; try one of these:</p>
+          <p>That URL doesn&apos;t exist on this site. The docs were recently reorganized; try one of these:</p>
           <ul>
             <li><a href="/docs/introduction">Nebari Kubernetes Platform (NKP) overview</a></li>
             <li><a href="/classic/welcome">Nebari Classic documentation</a></li>
