+++ docs/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@docusaurus/tsconfig",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "strict": true
+  },
+  "exclude": ["node_modules", "build", ".docusaurus"]
+}
