+++ netlify.toml
@@ -1,6 +1,6 @@
 [build]
 base = "docs/"
-command = "yarn run build"
+command = "npm run build"
 publish = "build/"
 
 # The default HTTP status code is 301, but you can define a different one.
